+++ spec/teamtailor/parser/job_application_spec.rb
@@ -14,18 +14,18 @@ RSpec.describe Teamtailor::JobApplication do
           json_payload.dig('included')
         )
         expect(job_application.candidate).to be_loaded
-        expect(job_application.candidate.record.id).to eq 410
+        expect(job_application.candidate.records.first.id).to eq 410
         expect(job_application.job).to be_loaded
-        expect(job_application.job.record.id).to eq 26
+        expect(job_application.job.records.first.id).to eq 26
 
         deserialized_job_application =
           Teamtailor::JobApplication.deserialize(job_application.serialize)
         expect(job_application.payload).to eq deserialized_job_application.payload
 
         expect(deserialized_job_application.candidate).to be_loaded
-        expect(deserialized_job_application.candidate.record.id).to eq 410
+        expect(deserialized_job_application.candidate.records.first.id).to eq 410
         expect(deserialized_job_application.job).to be_loaded
-        expect(deserialized_job_application.job.record.id).to eq 26
+        expect(deserialized_job_application.job.records.first.id).to eq 26
       end
     end
   end