+++ spec/teamtailor/record_spec.rb
@@ -84,7 +84,7 @@ RSpec.describe Teamtailor::Record do
       expect(record.value).to eq 42
       relation = record.user
       expect(relation).to be_loaded
-      expect(relation.record.name).to eq 'Marshall Mathers'
+      expect(relation.records.first.name).to eq 'Marshall Mathers'
     end
 
     describe 'serializing' do
@@ -97,18 +97,18 @@ RSpec.describe Teamtailor::Record do
           json_payload.dig('included')
         )
         expect(job_application.candidate).to be_loaded
-        expect(job_application.candidate.record.id).to eq 410
+        expect(job_application.candidate.records.first.id).to eq 410
         expect(job_application.job).to be_loaded
-        expect(job_application.job.record.id).to eq 26
+        expect(job_application.job.records.first.id).to eq 26
 
         deserialized_job_application =
           Teamtailor::Record.deserialize(job_application.serialize)
         expect(job_application.payload).to eq deserialized_job_application.payload
 
         expect(deserialized_job_application.candidate).to be_loaded
-        expect(deserialized_job_application.candidate.record.id).to eq 410
+        expect(deserialized_job_application.candidate.records.first.id).to eq 410
         expect(deserialized_job_application.job).to be_loaded
-        expect(deserialized_job_application.job.record.id).to eq 26
+        expect(deserialized_job_application.job.records.first.id).to eq 26
       end
     end
   end