+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- [breaking] Update `Teamtailor::Relationship` to support returning multiple records
 - Add `Teamtailor::Requisition` and `Client#requisitions`
 
 ## v0.2.6 - 2020-05-18